// src/api/orval/model/downloadFileRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export interface DownloadFileRequest {
  /**
   * @minimum 1
   * @maximum 2147483647
   */
  issuedOvpnFileId: number;
  /**
   * @minimum 1
   * @maximum 2147483647
   */
  vpnServerId: number;
}