// src/api/orval/model/downloadFileResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { IssuedOvpnFileDto } from "./issuedOvpnFileDto";

export interface DownloadFileResponse {
  issuedOvpn?: IssuedOvpnFileDto;
  fileSizeBytes?: number;
  /** @nullable */
  content?: string | null;
}