// src/api/orval/model/downloadFileResponseApiResponse.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { DownloadFileResponse } from "./downloadFileResponse";

export interface DownloadFileResponseApiResponse {
  success?: boolean;
  /** @nullable */
  message?: string | null;
  data?: DownloadFileResponse;
}